# Speed up `add_weights!`