Speed up `add_weights!` by only having vectors of the indices and weight value, and calculating everything else within the `for` loop

Old

```
julia> @btime LightOSM.add_weights!($g, :distance);
  859.221 ms (115 allocations: 240.18 MiB)

julia> @btime LightOSM.add_weights!($g, :time);
  1.759 s (136 allocations: 257.18 MiB)
```

New
```
julia> @btime LightOSM.add_weights!($g, :distance);
  720.114 ms (30 allocations: 118.47 MiB)

julia> @btime LightOSM.add_weights!($g, :time);
  1.335 s (30 allocations: 118.47 MiB)
```